Restore lr_wrapper into deployment script
