--- scripts/lr_wrapper.py
@@ -1,8 +1,0 @@
-import mlflow
-
-class lr_wrapper(mlflow.pyfunc.PythonModel):
-    def __init__(self, model):
-        self.model = model
-    
-    def predict(self, context, model_input):
-        return self.model.predict_proba(model_input)[:, 1]
